+++ csv/stringify_stream_test.ts
@@ -1,5 +1,5 @@
 // Copyright 2018-2024 the Deno authors. All rights reserved. MIT license.
-import { CsvStringifyStream } from "./csv_stringify_stream.ts";
+import { CsvStringifyStream } from "./stringify_stream.ts";
 import { StringifyError } from "./stringify.ts";
 import { assertEquals, assertRejects } from "@std/assert";
